requirements.txt
APIFlask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.20.0
pylint_flask_sqlalchemy==0.2.0
isort==5.10.1
black==22.6.0
pytest==7.1.2
coverage==6.4.3
PyMySQL==1.0.2
Faker==14.2.0
